K97: Meadows and Wheels complete. ArbWheelZ with bias and first TestSuite instance hooked up: CarlstromWheelZSpec.

+++ kosmos-core/src/main/kotlin/org/vorpal/kosmos/core/relations/instances/IntRelations.kt
@@ -6,23 +6,24 @@ import org.vorpal.kosmos.core.relations.TotalOrder
 import org.vorpal.kosmos.core.relations.TotalStrictOrder
 import org.vorpal.kosmos.core.relations.eqRelation
 import org.vorpal.kosmos.core.relations.ltRelation
+import java.math.BigInteger
 
 object IntRelations {
-    val IntComparator: Comparator<Int> =
+    val IntComparator: Comparator<BigInteger> =
         Comparator.naturalOrder()
 
-    val IntLT: Relation<Int> =
+    val IntLT: Relation<BigInteger> =
         IntComparator.ltRelation()
 
-    val IntTotalStrict: TotalStrictOrder<Int> =
+    val IntTotalStrict: TotalStrictOrder<BigInteger> =
         TotalStrictOrder.of(IntLT)
 
-    val IntEqByComparator: Relation<Int> =
+    val IntEqByComparator: Relation<BigInteger> =
         IntComparator.eqRelation(Symbols.EQUALS)
 
-    val IntLE: Relation<Int> =
+    val IntLE: Relation<BigInteger> =
         IntTotalStrict.leFrom(IntEqByComparator)
 
-    val IntTotalOrder: TotalOrder<Int> =
+    val IntTotalOrder: TotalOrder<BigInteger> =
         TotalOrder.of(IntLE)
 }
